Refatoração: PilhaDeCarne.cs
Renomeado agora para FleshStack.cs

Referências a esta classe foram corrigidas onde necessárias.

BUG INTRODUZIDO (do último commit): inimigos passam por cima da UI. Relacionado às sorting layers, fácil de arrumar.

+++ Assets/Scripts/FleshDrop.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 public class FleshDrop : MonoBehaviour, IClickable {
     private SpriteRenderer spriteRenderer;
     private BoxCollider2D boxCollider;
-    public float rottenPointsAmount;
+    public int rottenPointsAmount;
 
-    public static event Action<float> OnFleshClicked;
+    public static event Action<int> OnFleshClicked;
 
     public void OnClicked() {
         OnFleshClicked?.Invoke(rottenPointsAmount);
